requirements.txt
# Video2Text (V2T) - Python依赖清单
# 安装命令: pip install -r requirements.txt

# =============================================================================
# 核心依赖 - 必需
# =============================================================================

# OpenAI Whisper - 语音识别核心库
# 功能: 加载Whisper模型，执行语音转文本
# 模型文件会自动下载到 ~/.cache/whisper/ 目录
openai-whisper>=20231117

# FFmpeg-Python - FFmpeg的Python封装
# 功能: 调用系统FFmpeg从视频提取音频
# 注意: 此包仅提供Python绑定，系统仍需安装FFmpeg可执行文件
ffmpeg-python>=0.2.0

# tqdm - 进度条显示
# 功能: 显示音频提取和转录进度
# 支持: 终端进度条、预估剩余时间
tqdm>=4.65.0

# =============================================================================
# 可选依赖 - 增强功能
# =============================================================================

# NumPy - 数值计算库
# 用途: 音频数据处理、特征提取
# 注意: Whisper已依赖此包，此处显式列出以明确依赖关系
numpy>=1.24.0

# PyDub - 音频处理库
# 用途: 音频格式转换、切片、音量调整
# 适用场景: 长视频分段处理时的音频操作
# pydub>=0.25.1

# psutil - 系统监控库
# 用途: 监控内存使用、磁盘空间检查
# 适用场景: 大文件处理前的资源检查
# psutil>=5.9.0

# =============================================================================
# 开发依赖 - 仅开发环境需要
# =============================================================================

# pytest - 测试框架
# pytest>=7.4.0

# black - 代码格式化
# black>=23.0.0

# flake8 - 代码风格检查
# flake8>=6.0.0
